feat: break out form and add button
